backend/face_recognition_service/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy>=1.24
pillow
python-multipart
requests